feat: update security configuration and enhance token handling
