+++ src/main/java/com/burntoburn/easyshift/controller/TestController.java
@@ -3,7 +3,6 @@ package com.burntoburn.easyshift.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
@@ -19,12 +18,6 @@ public class TestController {
     @GetMapping("/api/test")
     public ResponseEntity<TestResponse> testEndpoint() {
         return ResponseEntity.ok(new TestResponse("테스트 성공"));
-    }
-
-    @GetMapping("/oauth2/code/kakao")
-    public String signinKakao(@RequestParam String code) {
-        System.out.println("code = " + code);
-        return "test";
     }
     
     public record TestResponse(String message) {}
