+++ src/main/java/com/burntoburn/easyshift/dto/token/CreateAccessTokenRequest.java
@@ -6,5 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CreateAccessTokenRequest {
+
     private String refreshToken;
+
 }
