+++ src/main/java/com/burntoburn/easyshift/dto/token/CreateAccessTokenResponse.java
@@ -8,4 +8,5 @@ import lombok.Getter;
 public class CreateAccessTokenResponse {
 
     private String accessToken;
+
 }
